refactor(admin-form.controller): shifts over handler for count retrieval from sms to admin-form
