+++ src/types/api/index.ts
@@ -7,4 +7,3 @@ export * from './examples'
 export * from './encrypt_submission'
 export * from './email_submission'
 export * from './submission'
-export * from './sms'
