--- src/types/api/sms.ts
@@ -1,3 +1,0 @@
-export interface SmsCountsDto {
-  smsCounts: number
-}
